Extract per-purpose entrypoints (pr/manual/local)
